requirements.txt
APScheduler==3.11.0
boto3==1.36.9
grpcio==1.67.1
grpcio_tools==1.67.1
langdetect==1.0.9
numpy==2.2.2
openai==1.61.1
Pillow==11.1.0
promptgen==0.0.2
protobuf==5.26.1
py_trees==2.3.0
PyAudio==0.2.14
pydub==0.25.1
python-dotenv==1.0.1
python-telegram-bot==21.10
pytz==2024.1
Requests==2.32.3
setuptools==75.1.0
soundfile==0.13.1
